% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_segment.R
\name{base_segment}
\alias{base_segment}
\title{Base arguments for segment function}
\usage{
base_segment(data, log_likelihood, penalty, initial_position,
  allow_parallel)
}
\arguments{
\item{data}{matrix for which we wish to estimate the independent segments of.}

\item{log_likelihood}{log likelihood estimation funciton, which will be
applied to all possible combinations of segments. Because it's executed
many times, it's likely to be the slow part of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant \code{multivariate} estimation.}

\item{penalty}{a function that determines the penalty for the segment. It's
called with the segment being analysed as it's only parameter.}

\item{initial_position}{a initial position for the recursive algorithm.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Defaults to TRUE.}
}
\description{
Describe base arguments for segment function
}
