% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berlin.R
\docType{data}
\name{berlin}
\alias{berlin}
\title{Daily temperatures from weather stations in Berlin}
\format{A matrix containing daily temperatures, with each column representing
a date and each column representing a weather station in Berlin
\describe{
\item{rows}{}
\item{columns}{dates from the years 2010 and 2011}
...
}}
\source{
\url{https://www.dwd.de/DE/Home/home_node.html}
}
\usage{
berlin
}
\description{
Contains weather daily weather data from many \href{https://dwd.de/}{Deutscher Wetterdienst}
weather stations in Berlin from the years of 2010 and 2011. Data was obtained using the
package \href{https://CRAN.R-project.org/package=rdwd}{rdwd}
and reformatted to a format appropriate to be used for analysis in this object.
}
\keyword{datasets}
