% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounded_sam.R
\name{custom_bbox_segmentation_cli}
\alias{custom_bbox_segmentation_cli}
\title{Perform Custom Bounded Box Segmentation}
\usage{
custom_bbox_segmentation_cli(
  image_path,
  bbox,
  polygon_refinement = FALSE,
  segmenter_id = "Zigeng/SlimSAM-uniform-77",
  script_path = system.file("python/main.py", package = "SegmentR"),
  output_plot = NULL,
  output_json = NULL,
  conda_env = "segmentr-env"
)
}
\arguments{
\item{image_path}{Character string. Path to the input image.}

\item{bbox}{Numeric vector. Custom bounding box coordinates = xmin, ymin, xmax, ymax.}

\item{polygon_refinement}{Logical. Whether to refine polygons (default: FALSE).}

\item{segmenter_id}{Character string. ID of the segmenter model.}

\item{script_path}{Character string. Path to the Python script.}

\item{output_plot}{Character string or NULL. Path to save the output plot. If NULL, saves to current working directory.}

\item{output_json}{Character string or NULL. Path to save the output JSON. If NULL, saves to current working directory.}

\item{conda_env}{Character string. Name of the conda environment to use.}
}
\value{
A list containing the segmentation results.
}
\description{
This function performs segmentation on an image using a custom bounding box.
}
