% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_border.R
\name{draw_border}
\alias{draw_border}
\title{Draw Border on Segmented Image}
\usage{
draw_border(segmented_array, edge_color = "black")
}
\arguments{
\item{segmented_array}{A numeric array representing the segmented image}

\item{edge_color}{A character string specifying the color of the edges (default is "black")}
}
\value{
A magick-image object with borders drawn
}
\description{
This function draws borders on a segmented image by detecting edges between different segments and allows the user to specify the border color.
}
\details{
The function takes a numeric array representing a segmented image and detects edges between segments. It then draws borders along these edges by setting the edge pixels to a specified color. The resulting image is converted back to a magick-image object and returned.
}
