% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{exclude_masks}
\alias{exclude_masks}
\title{Exclude masks from a combined mask}
\usage{
exclude_masks(combined_mask, exclude_masks)
}
\arguments{
\item{combined_mask}{Logical matrix. The combined mask to exclude from.}

\item{exclude_masks}{List of logical matrices representing masks to exclude.}
}
\value{
A logical matrix representing the resulting mask after exclusions.
}
\description{
This function excludes specified masks from a combined mask.
}
\examples{
\dontrun{
final_mask <- exclude_masks(combined_mask, bee_masks)
}
}
