% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_coco_json}
\alias{export_coco_json}
\title{Export segmentation in COCO JSON format}
\usage{
export_coco_json(segmentation, output_path)
}
\arguments{
\item{segmentation}{Result from load_segmentation_results}

\item{output_path}{Path to save the JSON file}
}
\value{
Invisibly returns path to saved file
}
\description{
Export segmentation in COCO JSON format
}
