% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_imagej_roi}
\alias{export_imagej_roi}
\title{Export segmentation as ImageJ ROIs}
\usage{
export_imagej_roi(segmentation, output_dir)
}
\arguments{
\item{segmentation}{Result from load_segmentation_results}

\item{output_dir}{Directory to save the ROIs}
}
\value{
Invisibly returns paths to saved files
}
\description{
Export segmentation as ImageJ ROIs
}
