% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_analysis.R
\name{extract_colors}
\alias{extract_colors}
\title{Extract color information from an image using a mask}
\usage{
extract_colors(image, mask, n_colors = 5)
}
\arguments{
\item{image}{Array. The original image (height x width x channels).}

\item{mask}{Logical matrix. The mask specifying which pixels to consider.}

\item{n_colors}{Integer. Number of dominant colors to extract (default: 5).}
}
\value{
A list containing color information including dominant colors and color statistics.
}
\description{
This function extracts color information from the parts of an image
specified by a mask.
}
\examples{
\dontrun{
color_info <- extract_colors(results$image, final_mask, n_colors = 3)
}
}
