% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_clusters.R
\name{find_optimal_clusters_combined}
\alias{find_optimal_clusters_combined}
\title{Find Optimal Number of Clusters}
\usage{
find_optimal_clusters_combined(img, max_clusters = 10)
}
\arguments{
\item{img}{An image object}

\item{max_clusters}{Maximum number of clusters to consider}
}
\value{
A list containing the optimal number of clusters and a plot
}
\description{
Find Optimal Number of Clusters
}
