% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounded_sam.R
\name{grounded_segmentation_r}
\alias{grounded_segmentation_r}
\title{Perform Grounded Segmentation}
\usage{
grounded_segmentation_r(
  image,
  labels,
  threshold = 0.3,
  polygon_refinement = FALSE,
  detector_id = NULL,
  segmenter_id = NULL
)
}
\arguments{
\item{image}{Path to image or URL}

\item{labels}{List of labels to detect}

\item{threshold}{Detection threshold}

\item{polygon_refinement}{Whether to refine polygons}

\item{detector_id}{ID of the detector model}

\item{segmenter_id}{ID of the segmenter model}
}
\value{
List containing image array and detections
}
\description{
Perform Grounded Segmentation
}
