% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_cols.R
\name{group_cols}
\alias{group_cols}
\title{group_cols}
\usage{
group_cols(hex_colors, n_clusters = 5)
}
\arguments{
\item{hex_colors}{A character vector of hex colour values (e.g., \code{"#FF0000"}).}

\item{n_clusters}{Integer. Number of clusters to form. Default is 5.}
}
\value{
A data frame with the original hex colours and their assigned group labels.
}
\description{
Group hex colours using k-means clustering in RGB space.
}
\examples{
hex_colors <- c("#FF0000", "#00FF00", "#0000FF", "#FFFF00", "#FF00FF")
group_cols(hex_colors, n_clusters = 2)

}
