% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridalg.R
\name{hybridalg}
\alias{hybridalg}
\title{Segment data into changepoints using a mixed hierarchical-exact approach}
\usage{
hybridalg(data, log_likelihood = multivariate, penalty = function(x) 0,
  allow_parallel = TRUE, max_segments = ncol(data), threshold = 50)
}
\arguments{
\item{data}{matrix for which we wish to estimate the independent segments of.}

\item{log_likelihood}{log likelihood estimation funciton, which will be
applied to all possible combinations of segments. Because it's executed
many times, it's likely to be the slow part of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant \code{multivariate} estimation.}

\item{penalty}{a function that determines the penalty for the segment. It's
called with the segment being analysed as it's only parameter.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Defaults to TRUE.}

\item{threshold}{the threshold for which the exact algorithm will be used,
i.e. when the number of columns in the segment is less than or equal to the
threshold.}
}
\description{
For the larger datasets, assume the data is hierarchical, but calculate
the exact segments when they're smaller than a threshold
}
\details{
This algorithm implements an approach mixing the hierarchical and exact
algorithms. It uses the hierarchical algorithms when the size of the segment
is bigger than the threshold, and then goes on to use the exact algorithm
when the size of the segment is less than or equal to the threshold.
}
