% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_to_palette.R
\name{img_to_palette}
\alias{img_to_palette}
\title{img_to_palette}
\usage{
img_to_palette(
  img,
  n = NULL,
  avg_cols = FALSE,
  exclude = FALSE,
  n_clusters = NULL,
  custom_exclude = NULL
)
}
\arguments{
\item{img}{magick image object \emph{or} path/URL to the image.}

\item{n}{Integer. Number of top colours to extract. Default is NULL (all).}

\item{avg_cols}{Logical. Whether to average colours into clusters. Default is FALSE.}

\item{exclude}{Logical. Whether to exclude near-black/white shades. Default is FALSE.}

\item{n_clusters}{Integer or NULL. Number of clusters to use for averaging. If NULL, clustering is skipped.}

\item{custom_exclude}{Optional character vector of hex colours to exclude.}
}
\value{
A character vector of hex codes representing the palette.
}
\description{
Generate a colour palette from an image based on the most frequent colours.
}
