% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loading.R
\name{load_image}
\alias{load_image}
\title{Load an image file}
\usage{
load_image(image_path)
}
\arguments{
\item{image_path}{Character string. Path to the input image file.}
}
\value{
A 3D array representing the image (height x width x color channels).
}
\description{
This function loads an image file and returns it as an array.
}
\examples{
\dontrun{
img <- load_image("path/to/image.jpg")
}
}
