% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loading.R
\name{load_segcolr_example_data}
\alias{load_segcolr_example_data}
\title{Load the SegColR example dataset}
\usage{
load_segcolr_example_data()
}
\value{
A list containing the example images, their file paths, and photographer credits.
}
\description{
Load the SegColR example dataset
}
