% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loading.R
\name{load_segmentr_example_data}
\alias{load_segmentr_example_data}
\title{Load the SegmentR example dataset}
\usage{
load_segmentr_example_data()
}
\value{
A list containing the example images, their file paths, and photographer credits.
}
\description{
Load the SegmentR example dataset
}
