% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-segment.R
\name{multivariate}
\alias{multivariate}
\title{Efficient Discrete Maximum Likelihood estimation function.}
\usage{
multivariate(X, na_action = na.omit)
}
\arguments{
\item{X}{a matrix to estimate the multivariate of. Each row is considered
to be an observation, and each column is considered to be a different
variable.}

\item{na_action}{A function that is applied to the `X` parameter. Defaults to `na.omit` function.}
}
\value{
the estimate of the Discrite Maximum Likelyhood for the dataframe provided.
}
\description{
Efficient Discrete Maximum Likelihood estimation function.
}
