% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_pixels_by_cluster.R
\name{plot.segmentR_raster}
\alias{plot.segmentR_raster}
\title{Plot a \code{segmentR_raster} object}
\usage{
\method{plot}{segmentR_raster}(x, col = hcl.colors(max(x$data)), axes = FALSE, square = TRUE, ...)
}
\arguments{
\item{x}{A \code{segmentR_raster}.}

\item{col}{Vector of colours for the classes (defaults to \code{hcl.colors()}).}

\item{axes}{Logical; draw axes?  Default \code{FALSE}.}

\item{square}{Logical; \code{TRUE} (default) draws each cell square;
\code{FALSE} preserves the overall raster proportions.}

\item{...}{Further arguments passed to \code{graphics::image()}.}
}
\description{
Plot a \code{segmentR_raster} object
}
