% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounded_sam.R
\name{plot_detections_plotly_r}
\alias{plot_detections_plotly_r}
\title{Plot Detections using Plotly}
\usage{
plot_detections_plotly_r(image_array, detections)
}
\arguments{
\item{image_array}{Image as a numpy array}

\item{detections}{List of detections}
}
\description{
Plot Detections using Plotly
}
