% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounded_sam.R
\name{plot_detections_r}
\alias{plot_detections_r}
\title{Plot Detections}
\usage{
plot_detections_r(image_array, detections, save_name = NULL)
}
\arguments{
\item{image_array}{Image as a numpy array}

\item{detections}{List of detections}

\item{save_name}{Optional name to save the plot}
}
\description{
Plot Detections
}
