% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_palette.R
\name{plot_palette}
\alias{plot_palette}
\title{plot_palette}
\usage{
plot_palette(x, color_col = NULL)
}
\arguments{
\item{x}{A character vector of hex colours, or a data frame.}

\item{color_col}{Optional. If \code{x} is a data frame, name of the column containing colours.}
}
\value{
A \code{ggplot2} plot object displaying the colour palette.
}
\description{
Plot a colour palette from a character vector or a data frame column of hex codes.
}
\examples{
# From a vector
plot_palette(c("#FF0000", "#00FF00", "#0000FF"))

# From a data frame
df <- data.frame(col = c("#FF0000", "#00FF00", "#0000FF"))
plot_palette(df, "col")

}
