% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting.R
\name{plot_repainted_mask}
\alias{plot_repainted_mask}
\title{Repaint Masked Image with Dominant Colors}
\usage{
plot_repainted_mask(image, final_mask, color_info)
}
\arguments{
\item{image}{Array. The original image (height x width x channels).}

\item{final_mask}{Logical matrix. The final mask specifying which pixels to consider.}

\item{color_info}{List. The dominant color information, including kmeans result.}
}
\value{
Array. The repainted image (height x width x channels).
}
\description{
This function takes an image, a final mask, and the dominant color information
and generates a new "repainted" image where the masked regions are colored
according to the dominant colors.
}
