% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_segmentr.R
\name{print.segmentr}
\alias{print.segmentr}
\title{Print a segmentr object}
\usage{
\method{print}{segmentr}(results)
}
\description{
Prints a short description of the segments found in the \code{segmentr} object
}
\details{
A short representation of the segments is printed on the screen, using the
\code{start:end} range notation.
}
