% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-segment.R
\name{r_multivariate}
\alias{r_multivariate}
\title{Estimated Discrete Maximum Likelihood implemented in R, used to benchmark
agaisnt the Rcpp implementation.}
\usage{
r_multivariate(X, na.omit = TRUE)
}
\arguments{
\item{X}{a matrix to estimate the multivariate of. Each row is considered
to be an observation, and each column is considered to be a different
variable.}

\item{na.omit}{If true, omits na from the dataset.}
}
\value{
the estimate of the Discrite Maximum Likelyhood for the dataframe provided.
}
\description{
Estimated Discrete Maximum Likelihood implemented in R, used to benchmark
agaisnt the Rcpp implementation.
}
