% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image.R
\name{read_image}
\alias{read_image}
\title{read_image}
\usage{
read_image(file_path, width = NULL, height = NULL, rotate = NULL)
}
\arguments{
\item{file_path}{Character. Path or URL to the image file.}

\item{width}{Integer or NULL. New width in pixels. Default is NULL.}

\item{height}{Integer or NULL. New height in pixels. Default is NULL.}

\item{rotate}{Numeric or NULL. Degrees to rotate the image clockwise. Default is NULL.}
}
\value{
A \code{magick-image} object.
}
\description{
This function reads an image from a specified file path (or URL) and optionally resizes it to the given width/height,
and/or rotates it by the specified number of degrees.
}
\details{
Read and optionally resize or rotate an image using the \code{magick} package.

Resizing maintains aspect ratio unless both dimensions are specified. Use \code{rotate} to rotate the image clockwise
(e.g., 90, 180, 270).
}
\examples{
\dontrun{
img <- read_image("https://upload.wikimedia.org/wikipedia/en/a/a1/Radiohead.okcomputer.albumart.jpg", width = 300)
plot(img)
}

}
