% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_image_from_url.R
\name{read_image_from_url}
\alias{read_image_from_url}
\title{read_image_from_url}
\usage{
read_image_from_url(path)
}
\arguments{
\item{path}{A character string representing the URL of the image file.}
}
\value{
An object containing the image data. If the image is a JPG, the object will be of class "array".
If the image is a PNG, the object will be of class "matrix".
}
\description{
Read a PNG or JPG image from a URL and return the image data.
}
\examples{
\dontrun{
img <- read_image_from_url("https://upload.wikimedia.org/wikipedia/en/0/02/Homer_Simpson_2006.png")
img
}

}
