% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grounded_sam.R
\name{run_grounded_segmentation}
\alias{run_grounded_segmentation}
\title{Run grounded segmentation on an image or folder of images}
\usage{
run_grounded_segmentation(
  path,
  labels,
  threshold = 0.3,
  detector_id = "IDEA-Research/grounding-dino-tiny",
  segmenter_id = "Zigeng/SlimSAM-uniform-77",
  output_plot = NULL,
  output_json = NULL,
  show_plot = FALSE,
  create_dir = TRUE,
  conda_env = "segmentr-env",
  pattern = "\\\\.(jpg|jpeg|png)$",
  recursive = FALSE
)
}
\arguments{
\item{path}{Character string. Path to an input image or directory containing images.}

\item{labels}{Character vector. Labels to detect in the image.}

\item{threshold}{Numeric. Detection threshold (default: 0.3).}

\item{detector_id}{Character string. ID of the detector model.}

\item{segmenter_id}{Character string. ID of the segmenter model.}

\item{output_plot}{Character string. Path or directory to save the output plot.}

\item{output_json}{Character string. Path or directory to save the output JSON.}

\item{show_plot}{Boolean. Whether a plot should automatically be shown after a segmentation.}

\item{create_dir}{Boolean. If the output directory doesn't exist, one will be created.}

\item{conda_env}{Character string. Name of the conda environment to use.}

\item{pattern}{Character string. File pattern to match when path is a directory (default: "\\.(jpg|jpeg|png)$").}

\item{recursive}{Boolean. Whether to search for images recursively in subdirectories (default: FALSE).}
}
\value{
A list containing results for each processed image.
}
\description{
Run grounded segmentation on an image or folder of images
}
