% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_check.R
\name{search_conda_locations}
\alias{search_conda_locations}
\title{Search for Conda Executable in Common Locations}
\usage{
search_conda_locations()
}
\value{
A character string representing the path to the Conda executable if found,
or NULL if Conda is not found in the specified locations.
}
\description{
This function searches for the Conda executable in common installation directories.
It looks in the user's home directory, as well as standard system-wide directories.
}
