% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{segment}
\usage{
segment(..., algorithm = "exact")
}
\arguments{
\item{...}{params to be passed to the underlying function.}

\item{algorithm}{can be of type `exact`, `hierarchical` or `hybrid`, Default: 'exact'}
}
\value{
returns an object of type `segmentr`
}
\description{
Segments data into changepoints using a specified algorithm
}
\details{
Uses the specified algorithm to segment data into changepoints.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link{exactalg}} for the exact algorithm, \code{\link{hieralg}} for the
hierarchical algorithm implementation, \code{\link{hybrid}} for the hybrid algorithm
implementation.
}
