% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_image.R
\name{segment_image}
\alias{segment_image}
\title{Segment Image Colors}
\usage{
segment_image(img, n_colors = 5)
}
\arguments{
\item{img}{An image object}

\item{n_colors}{Number of colors to segment into}
}
\value{
A segmented image as a numeric array
}
\description{
This function segments an image into a specified number of colors using k-means clustering.
}
\details{
The function takes an image object and converts it into a matrix of RGB values.
It then applies k-means clustering to group the colours into the specified number of clusters.
Each pixel in the image is replaced with the color of its corresponding cluster center. T
he segmented image is returned as a numeric array with the same dimensions as the original image.
}
