% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_likelihood.R, R/segmentr-deprecated.R
\name{calculate_likelihood}
\alias{calculate_likelihood}
\alias{segmentr-deprecated}
\title{Deprecated functions in package \pkg{segmentr}.}
\description{
The functions listed below are deprecated and will be defunct in
the near future. When possible, alternative functions with similar
functionality are also mentioned. Help pages for deprecated functions are
available at \code{help("-deprecated")}.
}
\section{\code{calculate_likelihood}}{

\code{calculate_likelihood} is deprecated. It's use turned out to not be practical during research
in this package's development.
}

\keyword{internal}
