% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/white_balance_auto.R
\name{white_balance_auto}
\alias{white_balance_auto}
\title{white_balance_auto}
\usage{
white_balance_auto(
  img,
  method = c("greyworld", "percentile"),
  lower = 0.01,
  upper = 0.99
)
}
\arguments{
\item{img}{magick image object \emph{or} path/URL.}

\item{method}{"greyworld" (default) or "percentile".}

\item{lower, upper}{Quantiles for percentile stretch.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    magick image object.
}\if{html}{\out{</div>}}
}
\description{
Automatic grey-world or percentile white-balance correction.
}
