% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_segment.R
\name{base_segment}
\alias{base_segment}
\title{Base arguments for segment function}
\usage{
base_segment(data, likelihood, penalty, max_segments, allow_parallel)
}
\arguments{
\item{data}{matrix for which to find the changepooints}

\item{likelihood}{a function receives the segment matrix as argument
and returns a likelihood estimation. This function is used to calculate the
changepoints that maximize the total likelihood. Depending on the algorithm
being used, this function is likely to be executed many times, in which
case it's also likely to be the bottleneck of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant \code{multivariate} estimation.}

\item{penalty}{a function that receives the segment as parameter and returns
the penalty to the segment, directly subtracted from the likelihood
estimated by the other function. The idea if for this to be used as a way
of avoiding undesirable results, e.g. to avoid large segments by provind a
penalty function that penalizes big segments, of the other way arround, by
providing a penalty function that penalizes small segments.}

\item{max_segments}{an integer that defines the maximum amount of segments to
split the data into.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Assumes a cluster is registered with the \code{foreach}
package. Defaults to TRUE.}
}
\description{
Describe base arguments for segment function
}
\keyword{internal}
