% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactalg.R
\name{exactalg}
\alias{exactalg}
\title{Segment data into exact changepoints}
\usage{
exactalg(data, max_segments = ncol(data),
  log_likelihood = multivariate, penalty = function(data) 0,
  allow_parallel = TRUE)
}
\arguments{
\item{data}{matrix for which to find the changepooints}

\item{max_segments}{an integer that defines the maximum amount of segments to
split the data into.}

\item{log_likelihood}{a function receives the the segment matrix as argument
and returns a likelihood estimation. This function is used to calculate the
changepoints that maximize the total likelihood. Depending on the algorithm
being used, this function is likely to be executed many times, in which
case it's also likely to be the bottleneck of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant \code{multivariate} estimation.}

\item{penalty}{a function that receives the segment as parameter and returns
the penalty to the segment, directly subtracted from the likelihood
estimated by the other function. The idea if for this to be used as a way
of avoiding undesirable results, e.g. to avoid large segments by provind a
penalty function that penalizes big segments, of the other way arround, by
providing a penalty function that penalizes small segments.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Assumes a cluster is registered with the \code{foreach}
package. Defaults to TRUE.}
}
\description{
Find changespoints in data calculating the penalized likelihood for all possible
segment combinations
}
\details{
Function that implements the dynamic programming algorithm, with the intent
of finding points of independent changepoints for which the log likelihood
function is maximized. It analyzes all possible combinations, returning the
changepoints that are garanteed to segment the data matrix in the maximum
likelihood independent changepoints. Because it analyzes all possible combinations
of changepoionts, it has a quadratic algorithm complexity, meaning it works
in an acceptable computation time, whereas time increases quadratically,
being quite long for longer data sequences one wish to apply. For longer datasets,
the hierarchical algorithm might be more adequate.
}
