% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactalg.R
\name{exactalg}
\alias{exactalg}
\title{Thorough segment function}
\usage{
exactalg(data, max_segments = ncol(data),
  log_likelihood = multivariate, penalty = function(data) 0,
  allow_parallel = TRUE)
}
\arguments{
\item{data}{A matrix for which we wish to estimate the independent changepoints
of.}

\item{max_segments}{the max number of segments allowed to be found. Defaults
to the number of columns in `x`.}

\item{log_likelihood}{log likelihood estimation funciton, which will be
applied to all possible combinations of segments. Because it's executed
many times, it's likely to be the slow part of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant `multivariate` estimation.}

\item{penalty}{a function that determines the penalty for the segment. It's
called with the segment being analysed as it's only parameter.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Defaults to TRUE.}
}
\description{
Function that implements the dynamic programming algorithm, with the intent
of finding points of independent changepoints for which the log likelihood
function is maximized. It analyzes all possible combination, returning the
changepoints that are garanteed to segment the data matrix in the maximum
likelihood independent changepoints.
}
