% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_shades.R
\name{exclude_shades}
\alias{exclude_shades}
\title{exclude_shades}
\usage{
exclude_shades(data, threshold = 0.1, custom_exclude = NULL)
}
\arguments{
\item{data}{A data frame with a column \code{hex} containing HEX colours.}

\item{threshold}{Numeric (default 0.1). Distance from black or white to consider exclusion.}

\item{custom_exclude}{Optional character vector of custom HEX colours to exclude.}
}
\value{
Filtered data frame.
}
\description{
Exclude near-black and near-white colours based on RGB distance.
}
