% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_optimal_clusters.R
\name{find_optimal_clusters}
\alias{find_optimal_clusters}
\title{Find Optimal Number of Clusters}
\usage{
find_optimal_clusters(img, max_clusters = 10, method = "elbow")
}
\arguments{
\item{img}{An image object}

\item{max_clusters}{Maximum number of clusters to consider}

\item{method}{Method to use: "elbow" or "silhouette"}
}
\value{
A list containing the optimal number of clusters and a plot
}
\description{
This function finds the optimal number of clusters for segmenting an
image using either the elbow method or the silhouette method.
}
\details{
The function converts an image to a matrix of RGB values and calculates the
total within-cluster sum of squares (WSS) for different numbers of clusters.
Depending on the chosen method ("elbow" or "silhouette"), it determines the
optimal number of clusters. The elbow method identifies the point where the
WSS starts to decrease less sharply, while the silhouette method calculates
average silhouette scores for different cluster numbers.
}
