% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridalg.R
\name{hybridalg}
\alias{hybridalg}
\title{Segment data into changepoints using a mixed hierarchical-exact approach}
\usage{
hybridalg(data, likelihood = multivariate, penalty = function(x) 0,
  allow_parallel = TRUE, max_segments = ncol(data), threshold = 50)
}
\arguments{
\item{data}{matrix for which to find the changepooints}

\item{likelihood}{a function receives the segment matrix as argument
and returns a likelihood estimation. This function is used to calculate the
changepoints that maximize the total likelihood. Depending on the algorithm
being used, this function is likely to be executed many times, in which
case it's also likely to be the bottleneck of the function execution, so
it's advised that this function should have a performant, native
implementation. Defaults to a performant \code{multivariate} estimation.}

\item{penalty}{a function that receives the segment as parameter and returns
the penalty to the segment, directly subtracted from the likelihood
estimated by the other function. The idea if for this to be used as a way
of avoiding undesirable results, e.g. to avoid large segments by provind a
penalty function that penalizes big segments, of the other way arround, by
providing a penalty function that penalizes small segments.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Assumes a cluster is registered with the \code{foreach}
package. Defaults to TRUE.}

\item{max_segments}{an integer that defines the maximum amount of segments to
split the data into.}

\item{threshold}{the threshold for which the exact algorithm will be used,
i.e. when the number of columns in the segment is less than or equal to the
threshold.}
}
\description{
For the larger datasets, assume the data is hierarchical, but calculate
the exact segments when they're smaller than a threshold
}
\details{
This algorithm implements an approach mixing the hierarchical and exact
algorithms. It uses the hierarchical algorithms when the size of the segment
is bigger than the threshold, and then goes on to use the exact algorithm
when the size of the segment is less than or equal to the threshold.
}
