% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_multivariate.R
\name{r_multivariate}
\alias{r_multivariate}
\title{Logarithmic Discrete Multivariate Likelihood estimation function implemented
in R}
\usage{
r_multivariate(data, na.omit = TRUE)
}
\arguments{
\item{data}{Matrix to estimate the multivariate of. Each row is considered to
be an observation, and each column is considered to be a different
variable.}

\item{na.omit}{If true, omits NAs from the dataset.}
}
\value{
The estimate of the Discrete Maximum Likelihood for the dataframe
provided.
}
\description{
Estimate the likelihood of a given segment using the discrete multivariate
estimation, but code runs more slowly due to R implementation
}
\details{
This log likelihood function is implemented in R in order to be used to
benchmark against the \code{\link[=multivariate]{multivariate()}} version implemented in C++ for
performance.
}
