% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-segment.R
\name{segment}
\alias{segment}
\title{Function that implements the dynamic programming algorithm, with the intent
of finding points of independent segments for which the log likelihood
function is maximized.}
\usage{
segment(x, segmax = ncol(x), loglikfun = multivariate,
  penality = function(x) 0)
}
\arguments{
\item{x}{A matrix for which we wish to estimate the independent segments of.}

\item{loglikfun}{log likelihood estimation funciton, which will be applied to
all possible combinations of segments. Because it's executed many times,
it's likely to be the slow part of the function execution, so it's advised
that this function should have a performant, native implementation.
Defaults to a performant `multivariate` estimation.}

\item{penalty}{a function that determines the penalty for the segment. It's
called with the segment being analysed as it's only parameter.}
}
\description{
Function that implements the dynamic programming algorithm, with the intent
of finding points of independent segments for which the log likelihood
function is maximized.
}
