% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Function that implements the dynamic programming algorithm, with the intent
of finding points of independent segments for which the log likelihood
function is maximized.}
\usage{
segment(data, max_segments = ncol(data), log_likelihood = multivariate,
  penalty = function(data) 0, allow_parallel = TRUE)
}
\arguments{
\item{data}{A matrix for which we wish to estimate the independent segments of.}

\item{max_segments}{the max number of segments allowed to be found. Defaults
to the number of columns in `x`.}

\item{log_likelihood}{log likelihood estimation funciton, which will be applied to
all possible combinations of segments. Because it's executed many times,
it's likely to be the slow part of the function execution, so it's advised
that this function should have a performant, native implementation.
Defaults to a performant `multivariate` estimation.}

\item{penalty}{a function that determines the penalty for the segment. It's
called with the segment being analysed as it's only parameter.}

\item{allow_parallel}{allows parallel execution to take place using the
registered cluster. Defaults to TRUE.}
}
\description{
Function that implements the dynamic programming algorithm, with the intent
of finding points of independent segments for which the log likelihood
function is maximized.
}
