% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_image.R
\name{segment_image}
\alias{segment_image}
\title{segment_image}
\usage{
segment_image(img, n_colors = 5)
}
\arguments{
\item{img}{A \code{magick-image} object.}

\item{n_colors}{Integer ≥ 2. Number of colour clusters. Default is 5.}
}
\value{
A numeric array with the same dimensions as the input image,
where each pixel is assigned the RGB value of its cluster centroid.
}
\description{
Converts an image into an RGB matrix and uses k-means to cluster pixels into \code{n_colors} groups.
Each pixel is recoloured using its cluster centroid. The result is returned as a numeric RGB array.
}
\details{
Segment an image into a specified number of dominant colours using k-means clustering.
}
\examples{
img_path <- system.file("extdata", "sample_img.png", package = "segmentR")
img <- read_image(img_path, width = 300)
seg <- segment_image(img, n_colors = 2)
plot_segmented(img, seg)

}
